chore: resolve conflicts and move validate function to separate file
