+++ package.json
@@ -40,11 +40,7 @@
     "locutus": "2.0.16",
     "luxon": "2.3.2",
     "mysql": "2.18.1",
-<<<<<<< HEAD
     "nanoid": "3.3.4"
-=======
-    "nanoid": "^3.3.4"
->>>>>>> 80ea319 (feat(account deletion): add queueing functionality)
   },
   "devDependencies": {
     "@apollo/federation": "0.36.1",
