+++ src/server/main.ts
@@ -41,6 +41,7 @@ app.use(express.json());
 
 // Initialize routes
 app.use('/batchDelete', batchDeleteRouter);
+app.use('/queueDelete', queueDeleteRouter);
 
 // Initialize event handlers
 initItemEventHandlers(itemsEventEmitter, [
@@ -60,9 +61,6 @@ const server = startServer(contextFactory);
 
 app.use(xrayExpress.openSegment('list-api'));
 //Set XRay to use the host header to open its segment name.
-
-app.use(express.json());
-app.use('/queue-delete', queueDeleteRouter);
 
 AWSXRay.middleware.enableDynamicNaming('*');
 server.then((server) => server.applyMiddleware({ app, path: '/' }));
