+++ src/server/routes/batchDelete.ts
@@ -1,8 +1,9 @@
 import express from 'express';
-import { checkSchema, Schema, validationResult } from 'express-validator';
+import { checkSchema, Schema } from 'express-validator';
 import { SavedItemDataService } from '../../dataService';
 import { writeClient } from '../../database/client';
 import { nanoid } from 'nanoid';
+import { validate } from './validator';
 
 const router = express.Router();
 const batchDeleteSchema: Schema = {
@@ -43,21 +44,6 @@ const batchDeleteSchema: Schema = {
     },
   },
 };
-
-function validate(
-  req: express.Request,
-  res: express.Response,
-  next: express.NextFunction
-) {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    return res
-      .status(400)
-      .json({ errors: errors.array() })
-      .setHeader('Content-Type', 'application/json');
-  }
-  next();
-}
 
 router.post(
   '/',
